Split out NumericParts
