+++ core/src/ast/mod.rs
@@ -2,7 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::types::{Digits, Numeric};
+use crate::output::Digits;
+use crate::types::Numeric;
 use chrono_tz::Tz;
 use serde_derive::Serialize;
 use std::fmt;
